Refactor utility methods: formatDuration(), formatPeriod() and formatStatus()
