+++ az-core/src/main/java/azkaban/utils/TimeUtils.java
@@ -20,6 +20,8 @@ import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import org.joda.time.DurationFieldType;
+import org.joda.time.ReadablePeriod;
 
 public class TimeUtils {
 
@@ -48,6 +50,73 @@ public class TimeUtils {
     final ZonedDateTime zonedDateTime = ZonedDateTime.ofInstant(Instant.ofEpochMilli(timestampMs),
         ZoneId.systemDefault());
     return formatter.format(zonedDateTime);
+  }
+
+  public static String formatDuration(final long startTime, final long endTime) {
+    if (startTime == -1) {
+      return "-";
+    }
+
+    final long durationMS;
+    if (endTime == -1) {
+      durationMS = System.currentTimeMillis() - startTime;
+    } else {
+      durationMS = endTime - startTime;
+    }
+
+    long seconds = durationMS / 1000;
+    if (seconds < 60) {
+      return seconds + " sec";
+    }
+
+    long minutes = seconds / 60;
+    seconds %= 60;
+    if (minutes < 60) {
+      return minutes + "m " + seconds + "s";
+    }
+
+    long hours = minutes / 60;
+    minutes %= 60;
+    if (hours < 24) {
+      return hours + "h " + minutes + "m " + seconds + "s";
+    }
+
+    final long days = hours / 24;
+    hours %= 24;
+    return days + "d " + hours + "h " + minutes + "m";
+  }
+
+  public static String formatPeriod(final ReadablePeriod period) {
+    String periodStr = "null";
+
+    if (period == null) {
+      return periodStr;
+    }
+
+    if (period.get(DurationFieldType.years()) > 0) {
+      final int years = period.get(DurationFieldType.years());
+      periodStr = years + " year(s)";
+    } else if (period.get(DurationFieldType.months()) > 0) {
+      final int months = period.get(DurationFieldType.months());
+      periodStr = months + " month(s)";
+    } else if (period.get(DurationFieldType.weeks()) > 0) {
+      final int weeks = period.get(DurationFieldType.weeks());
+      periodStr = weeks + " week(s)";
+    } else if (period.get(DurationFieldType.days()) > 0) {
+      final int days = period.get(DurationFieldType.days());
+      periodStr = days + " day(s)";
+    } else if (period.get(DurationFieldType.hours()) > 0) {
+      final int hours = period.get(DurationFieldType.hours());
+      periodStr = hours + " hour(s)";
+    } else if (period.get(DurationFieldType.minutes()) > 0) {
+      final int minutes = period.get(DurationFieldType.minutes());
+      periodStr = minutes + " minute(s)";
+    } else if (period.get(DurationFieldType.seconds()) > 0) {
+      final int seconds = period.get(DurationFieldType.seconds());
+      periodStr = seconds + " second(s)";
+    }
+
+    return periodStr;
   }
 
 }
