+++ az-core/src/main/java/azkaban/utils/Utils.java
@@ -38,7 +38,6 @@ import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
 import org.apache.commons.io.IOUtils;
 import org.apache.log4j.Logger;
-import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.joda.time.Days;
 import org.joda.time.DurationFieldType;
@@ -301,40 +300,6 @@ public class Utils {
     } catch (final InstantiationException e) {
       throw new IllegalStateException(e);
     }
-  }
-
-  public static String formatDuration(final long startTime, final long endTime) {
-    if (startTime == -1) {
-      return "-";
-    }
-
-    final long durationMS;
-    if (endTime == -1) {
-      durationMS = DateTime.now().getMillis() - startTime;
-    } else {
-      durationMS = endTime - startTime;
-    }
-
-    long seconds = durationMS / 1000;
-    if (seconds < 60) {
-      return seconds + " sec";
-    }
-
-    long minutes = seconds / 60;
-    seconds %= 60;
-    if (minutes < 60) {
-      return minutes + "m " + seconds + "s";
-    }
-
-    long hours = minutes / 60;
-    minutes %= 60;
-    if (hours < 24) {
-      return hours + "h " + minutes + "m " + seconds + "s";
-    }
-
-    final long days = hours / 24;
-    hours %= 24;
-    return days + "d " + hours + "h " + minutes + "m";
   }
 
   public static Object invokeStaticMethod(final ClassLoader loader, final String className,
