+++ azkaban-common/src/test/resources/azkaban/executor/mail/firstErrorMessage.html
@@ -8,11 +8,11 @@
   </tr>
   <tr>
     <td>End Time</td>
-    <td>-</td>
+    <td>2016/07/17 11:54:16 EEST</td>
   </tr>
   <tr>
     <td>Duration</td>
-    <td>10 sec</td>
+    <td>5 sec</td>
   </tr>
   <tr>
     <td>Status</td>
