+++ azkaban-web-server/src/main/resources/azkaban/webapp/servlet/velocity/executingflowtriggerspage.vm
@@ -159,10 +159,12 @@
               #end
               <td>${trigger.getStatus()}</td>
               #if (${trigger.getStatus()} != "RUNNING" && ${trigger.getStatus()} != "CANCELLING")
-                <td>$utils.formatDuration(${trigger.getStartTime()}, ${trigger.getEndTime()})
+                <td>
+                  $TimeUtils.formatDuration(${trigger.getStartTime()}, ${trigger.getEndTime()})
                 </td>
               #else
-                <td>$utils.formatDuration(${trigger.getStartTime()}, ${System.currentTimeMillis()})
+                <td>
+                  $TimeUtils.formatDuration(${trigger.getStartTime()}, ${System.currentTimeMillis()})
                 </td>
               #end
 
@@ -213,10 +215,12 @@
                 #end
                 <td>${dep.getStatus()}</td>
                 #if (${dep.getEndTime()} != "0")
-                  <td>$utils.formatDuration(${dep.getStartTime()}, ${dep.getEndTime()})
+                  <td>
+                    $TimeUtils.formatDuration(${dep.getStartTime()}, ${dep.getEndTime()})
                   </td>
                 #else
-                  <td>$utils.formatDuration(${dep.getStartTime()}, ${System.currentTimeMillis()})
+                  <td>
+                    $TimeUtils.formatDuration(${dep.getStartTime()}, ${System.currentTimeMillis()})
                   </td>
                 #end
               </tr>
@@ -231,10 +235,12 @@
             <td>${flow.getFirst().proxyUsers}</td>
             <td>$TimeUtils.formatDateTime(${flow.getFirst().startTime})</td>
             <td>$TimeUtils.formatDateTime(${flow.getFirst().endTime})</td>
-            <td>$utils.formatDuration(${flow.getFirst().startTime}, ${flow.getFirst().endTime})</td>
+            <td>$TimeUtils.formatDuration(${flow.getFirst().startTime}, ${flow.getFirst().endTime})
+            </td>
             <td>
               <div
-                  class="status ${flow.getFirst().status}">$utils.formatStatus(${flow.getFirst().status})</div>
+                  class="status ${flow.getFirst().status}">$WebUtils.formatStatus(${flow.getFirst()
+                  .status})</div>
             </td>
             <td>
               <button type="button" id="cancelbtn" class="btn btn-danger btn-sm"
@@ -308,10 +314,12 @@
               #end
               <td>${trigger.getStatus()}</td>
               #if (${trigger.getStatus()} != "RUNNING" && ${trigger.getStatus()} != "CANCELLING")
-                <td>$utils.formatDuration(${trigger.getStartTime()}, ${trigger.getEndTime()})
+                <td>
+                  $TimeUtils.formatDuration(${trigger.getStartTime()}, ${trigger.getEndTime()})
                 </td>
               #else
-                <td>$utils.formatDuration(${trigger.getStartTime()}, ${System.currentTimeMillis()})
+                <td>
+                  $TimeUtils.formatDuration(${trigger.getStartTime()}, ${System.currentTimeMillis()})
                 </td>
               #end
               #if (${trigger.getFlowExecId()} == "-1")
@@ -358,10 +366,12 @@
                 <td>${dep.getStatus()}</td>
                 <td>${dep.getCancellationCause()}</td>
                 #if (${dep.getEndTime()} != "0")
-                  <td>$utils.formatDuration(${dep.getStartTime()}, ${dep.getEndTime()})
+                  <td>
+                    $TimeUtils.formatDuration(${dep.getStartTime()}, ${dep.getEndTime()})
                   </td>
                 #else
-                  <td>$utils.formatDuration(${dep.getStartTime()}, ${System.currentTimeMillis()})
+                  <td>
+                    $TimeUtils.formatDuration(${dep.getStartTime()}, ${System.currentTimeMillis()})
                   </td>
                 #end
               </tr>
