+++ azkaban-web-server/src/main/resources/azkaban/webapp/servlet/velocity/executionspage.vm
@@ -109,10 +109,13 @@
               <td>${flow.getFirst().proxyUsers}</td>
               <td>$TimeUtils.formatDateTime(${flow.getFirst().startTime})</td>
               <td>$TimeUtils.formatDateTime(${flow.getFirst().endTime})</td>
-              <td>$utils.formatDuration(${flow.getFirst().startTime}, ${flow.getFirst().endTime})</td>
               <td>
-                <div
-                    class="status ${flow.getFirst().status}">$utils.formatStatus(${flow.getFirst().status})</div>
+                $TimeUtils.formatDuration(${flow.getFirst().startTime}, ${flow.getFirst().endTime})
+              </td>
+              <td>
+                <div class="status ${flow.getFirst().status}">
+                  $WebUtils.formatStatus(${flow.getFirst().status})
+                </div>
               </td>
               <td>
                 <button type="button" id="cancelbtn" class="btn btn-danger btn-sm"
@@ -171,9 +174,9 @@
               <td>${flow.proxyUsers}</td>
               <td>$TimeUtils.formatDateTime(${flow.startTime})</td>
               <td>$TimeUtils.formatDateTime(${flow.endTime})</td>
-              <td>$utils.formatDuration(${flow.startTime}, ${flow.endTime})</td>
+              <td>$TimeUtils.formatDuration(${flow.startTime}, ${flow.endTime})</td>
               <td>
-                <div class="status ${flow.status}">$utils.formatStatus(${flow.status})</div>
+                <div class="status ${flow.status}">$WebUtils.formatStatus(${flow.status})</div>
               </td>
               <td></td>
             </tr>
