+++ azkaban-web-server/src/main/resources/azkaban/webapp/servlet/velocity/historypage.vm
@@ -117,9 +117,9 @@
                 <td>${flow.submitUser}</td>
                 <td>$TimeUtils.formatDateTime(${flow.startTime})</td>
                 <td>$TimeUtils.formatDateTime(${flow.endTime})</td>
-                <td>$utils.formatDuration(${flow.startTime}, ${flow.endTime})</td>
+                <td>$TimeUtils.formatDuration(${flow.startTime}, ${flow.endTime})</td>
                 <td>
-                  <div class="status ${flow.status}">$utils.formatStatus(${flow.status})</div>
+                  <div class="status ${flow.status}">$WebUtils.formatStatus(${flow.status})</div>
                 </td>
                 <td></td>
               </tr>
