+++ azkaban-web-server/src/main/resources/azkaban/webapp/servlet/velocity/jobhistorypage.vm
@@ -131,10 +131,10 @@
                 </td>
                 <td>$TimeUtils.formatDateTime(${job.startTime})</td>
                 <td>$TimeUtils.formatDateTime(${job.endTime})</td>
-                <td>$utils.formatDuration(${job.startTime}, ${job.endTime})</td>
+                <td>$TimeUtils.formatDuration(${job.startTime}, ${job.endTime})</td>
                 <td>
                   <div class="status ${job.status}">
-                    $utils.formatStatus(${job.status})
+                    $WebUtils.formatStatus(${job.status})
                   </div>
                 </td>
                 <td class="logLink">
